[refactor] image upload api 분리
