+++ src/app/(community)/(components)/PostForm.tsx
@@ -15,7 +15,14 @@ const PostForm = () => {
   const [content, setContent] = useState<string>('');
   const [category, setCategory] = useState('질문');
   const router = useRouter();
-
+  
+  const categories = [
+    { id: 'question', value: '질문', label: '질문' },
+    { id: 'chat', value: '잡담', label: '잡담' },
+    { id: 'study', value: '공부', label: '공부' },
+    { id: 'diary', value: '일기', label: '일기' }
+  ];
+  
   const {
     data: profile,
     isLoading,
@@ -80,12 +87,6 @@ const PostForm = () => {
     console.log(data);
   };
 
-  const categories = [
-    { id: 'question', value: '질문', label: '질문' },
-    { id: 'chat', value: '잡담', label: '잡담' },
-    { id: 'study', value: '공부', label: '공부' },
-    { id: 'diary', value: '일기', label: '일기' }
-  ];
 
   return (
     <form onSubmit={handleNewPost}>
